Refactor CatBin and OrdinalBin to share newVariableWindowProps
